improving models
add optional parameter for reading method, add method to get the readable data from a select order. 
add a method to get primay key (first index) and columun of a table to make an insert if id is not AI. 
prepareThenExecute now return a message if a request failed and boll(true) if the request is done.